🚚 유저 정보 접근 방법 설명하는 주석 signUp 페이지로 이동
